@@ -9,6 +9,7 @@ gem 'jquery-rails'
 gem 'turbolinks'
 gem 'jbuilder', '~> 1.2'
  # localart
+gem 'sorcery'
 gem 'activerecord-tableless'
 gem 'bootstrap-sass', '>= 3.0.0.0'
 gem 'figaro'
@@ -16,7 +17,7 @@ gem 'gibbon'
 gem 'high_voltage'
 gem 'simple_form', '~> 3.0.0.rc'
 gem 'pg'
-gem 'bcrypt-ruby', '~> 3.0.0', :require => "bcrypt"
+gem 'bcrypt-ruby', '~> 3.1.2', :require => "bcrypt"
 group :development do
 	gem 'better_errors'
 	gem 'quiet_assets' 
