@@ -1,29 +1,10 @@
 class Artist < ActiveRecord::Base
-  #attr_accessor :email, :farm, :name, :produce, :produce_price, :wepay_access_token, :wepay_account_id, :password
+  authenticates_with_sorcery!
 
-  validates :password, :presence => true
-  validates :password, :length => { :in => 6..200}
-  validates :name, :email, :presence => true
-  validates :email, :uniqueness => { :case_sensitive => false }
-  validates_format_of :email,
-		:with => /\A[-a-z0-9_+\.]+\@([-a-z0-9]+\.)+[a-z0-9]{2,4}\z/i
+  validates :password, length: { minimum: 3 }
+  validates :password, confirmation: true
+  validates :password_confirmation, presence: true
 
-  def password
-    password_hash ? @password ||= BCrypt::Password.new(password_hash) : nil
-  end
-
-  def password=(new_password)
-    @password = BCrypt::Password.create(new_password)
-    self.password_hash = @password
-  end
-
-def self.authenticate(email, test_password)
-	artist = Artist.find_by_email(email)
-	if artist && artist.password == test_password
-		artist
-	else
-		nil
-	end
-end
-
+  validates :email, uniqueness: true
+  
 end
